+++ packages/smarthr-ui/src/components/Dialog/ControlledFormDialog/stories/VRTFormDialog.stories.tsx
@@ -1,5 +1,5 @@
 import { Button } from '../../../Button'
-import { FormControl } from '../../../FormControl'
+import { FormControl } from '../../../FormGroup'
 import { Input } from '../../../Input'
 import { ControlledFormDialog } from '../ControlledFormDialog'
