+++ packages/smarthr-ui/src/components/Dialog/Dialog.test.tsx
@@ -5,8 +5,7 @@ import { type FC, useRef, useState } from 'react'
 import { IntlProvider } from '../../intl'
 import { Button } from '../Button'
 import { DatePicker } from '../DatePicker'
-import { Fieldset } from '../Fieldset'
-import { FormControl } from '../FormControl'
+import { Fieldset, FormControl } from '../FormGroup'
 import { Input } from '../Input'
 import { Cluster } from '../Layout'
 import { RadioButton } from '../RadioButton'
