+++ packages/smarthr-ui/src/components/Dropdown/SortDropdown/SortDropdown.tsx
@@ -17,8 +17,7 @@ import {
 import { useLatest } from '../../../hooks/useLatest'
 import { Localizer } from '../../../intl'
 import { Button } from '../../Button'
-import { Fieldset } from '../../Fieldset'
-import { FormControl } from '../../FormControl'
+import { Fieldset, FormControl } from '../../FormGroup'
 import { FaArrowDownWideShortIcon, FaArrowUpWideShortIcon } from '../../Icon'
 import { Cluster, Stack } from '../../Layout'
 import { RadioButton } from '../../RadioButton'
