--- packages/smarthr-ui/src/components/Fieldset/index.ts
@@ -1,1 +1,0 @@
-export { Fieldset } from './Fieldset'
