--- packages/smarthr-ui/src/components/FormControl/index.ts
@@ -1,1 +1,0 @@
-export { FormControl } from './FormControl'
