+++ packages/smarthr-ui/src/components/FormGroup/Fieldset.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen, waitFor } from '@testing-library/react'
 
-import { FormControl } from '../FormControl'
 import { Input } from '../Input'
 
 import { Fieldset } from './Fieldset'
+import { FormControl } from './FormControl'
 
 describe('Fieldset', () => {
   it('子要素が可視ラベルを持つフォームコントロール要素の場合、アクセシブルネームは可視ラベルになる', async () => {
