+++ packages/smarthr-ui/src/components/FormGroup/index.ts
@@ -1,1 +1,2 @@
-export { FormGroup } from './FormGroup'
+export { FormControl } from './FormControl'
+export { Fieldset } from './Fieldset'
